LayoutTests/fast/events/gesture/wheel-gesture-event-coordinates-in-iframe.html

<!DOCTYPE html>
<html>
<head>
<style>
body, html {
    width: 100%;
    height: 100%;
    margin: 0;
    padding; 0;
}

iframe {
    border: 1px solid red;
    width: 300px;
    height: 300px;
    margin: 0;
    padding: 0;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

const eventTypeWasObservedMap = new Map([
    ["wheel", false],
    ["gesturestart", false],
    ["gesturechange", false],
    ["gestureend", false],
]);

function eventTypeWasObserved(eventType) {
    eventTypeWasObservedMap.set(eventType, true);
}

function allEventsWereObserved() {
    let result = true;
    for (const [_, eventTypeWasObserved] of eventTypeWasObservedMap)
        result &&= eventTypeWasObserved;
    return result;
}

addEventListener("load", async () => {
    description(`This test verifies that gesture events (and wheel events generated from gestures) originating from an iframe have the correct coordinates relative to the iframe.`);

    const iframe = document.querySelector("iframe");
    const iframeStyle = window.getComputedStyle(iframe);
    const expectedEventPositionInIframe = {
        x: parseInt(iframeStyle.getPropertyValue("width"))/2,
        y: parseInt(iframeStyle.getPropertyValue("height"))/2
    };

    window.onmessage = (event) => {
        if (event.data.type === "iframeEvent") {
            const data = event.data.data;
            if (!eventTypeWasObservedMap.has(data.eventType))
                testFailed(`Unexpected event of type '${data.eventType}' observed.`);
            eventTypeWasObservedMap.set(data.eventType, true);
            if (data.clientX === expectedEventPositionInIframe.x && data.clientY === expectedEventPositionInIframe.y)
                testPassed(`${data.eventType}: 'event.clientX' is ${expectedEventPositionInIframe.x} and 'event.clientY' is ${expectedEventPositionInIframe.y}.`);
            else
                testFailed(`${data.eventType}: 'event.clientX' is ${data.clientX} (expected ${expectedEventPositionInIframe.x}) and 'event.clientY' is ${data.clientY} (expected ${expectedEventPositionInIframe.y}).`);
        }

        if (allEventsWereObserved())
            finishJSTest();
    };

    if (!window.eventSender) {
        debug("FAIL: This test requires window.eventSender.");
        return;
    }

    const middleOfIframeInWindow = {
        x: (iframe.offsetLeft + (iframe.offsetWidth / 2)),
        y: (iframe.offsetTop + (iframe.offsetHeight / 2))
    };
    await UIHelper.moveMouseAndWaitForFrame(middleOfIframeInWindow.x, middleOfIframeInWindow.y);
    eventSender.scaleGestureStart(0.01);
    eventSender.scaleGestureChange(0.01);
    eventSender.scaleGestureEnd(0.01);
});
</script>
</head>
<body>
    <iframe id="iframe" src="resources/gesture-iframe.html"></iframe>
</body>
</html>
