LayoutTests/fast/events/gesture/wheel-gesture-event-coordinates-in-iframe-expected.txt

This test verifies that gesture events (and wheel events generated from gestures) originating from an iframe have the correct coordinates relative to the iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gesturestart: 'event.clientX' is 150 and 'event.clientY' is 150.
PASS wheel: 'event.clientX' is 150 and 'event.clientY' is 150.
PASS gesturechange: 'event.clientX' is 150 and 'event.clientY' is 150.
PASS gestureend: 'event.clientX' is 150 and 'event.clientY' is 150.
PASS successfullyParsed is true

TEST COMPLETE
